+++ app/jobs/anomonitor/application_job.rb
@@ -1,4 +1,0 @@
-module Anomonitor
-  class ApplicationJob < ActiveJob::Base
-  end
-end